+++ lib/core/constants/app_constants.dart
@@ -1,7 +1,7 @@
 class AppConstants {
   AppConstants._();
 
-  static const String appName = 'Fbro';
+  static const String appName = 'DROP';
   static const String usersCollection = 'users';
   static const String shiftsCollection = 'shifts';
   static const String tasksCollection = 'tasks';